Fix cmesh readmshfile bug with orientation
For a face-to-face connection one face is declared
as the 'smaller face'. This info is used when the
orientation is computed.
For cmesh read msh file the computation of the smaller face
was wrong.
If both eclasses of the trees are the same then the 'smaller face'
is the face with smaller face_id.
Instead in readmshfile the face of the tree with smaller tree id
was used.

We fixed this behaviour.